GitHub actions build, unit and android tests
