+++ app/src/androidTest/java/com/missionchurchcooljc/mcc/ExampleInstrumentedTest.kt
@@ -1,14 +1,24 @@
+/*
+ * Copyright (c) 2021 Kevin Phillips, Mission Church of Our Lord Jesus Christ
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.missionchurchcooljc.mcc
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
-import com.google.gson.stream.JsonReader
-import com.missionchurchcooljc.data_android.WebsiteHighlight
-import com.missionchurchcooljc.mcc.utilities.ABOUT_US_DATA_FILENAME
-import org.hamcrest.CoreMatchers.instanceOf
-import org.junit.Assert.*
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -27,18 +37,5 @@ class ExampleInstrumentedTest {
         assertEquals("io.techministry.android.missionchurch", appContext.packageName)
     }
 
-    @Test
-    fun test_JsonReader() {
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        val f = appContext.assets.open(ABOUT_US_DATA_FILENAME)
-        val jsonReader = JsonReader(f.reader())
-        val whType = object : TypeToken<List<WebsiteHighlight>>() {}.type
-        val whList: List<WebsiteHighlight> = Gson().fromJson(jsonReader, whType)
-//        assertEquals(jsonReader, JsonReader(f.reader()))
-        whList.forEach { t ->
-            assertThat(t, instanceOf(WebsiteHighlight::class.java))
 
-
-        }
-    }
 }
