+++ app/src/androidTest/java/com/missionchurchcooljc/mcc/data/WebsiteHighlightDaoTest.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright 2018 Google LLC
+ * Copyright (c) 2021 Kevin Phillips, Mission Church of Our Lord Jesus Christ
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     https://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -25,9 +25,11 @@ import com.google.gson.reflect.TypeToken
 import com.google.gson.stream.JsonReader
 import com.missionchurchcooljc.data_android.WebsiteHighlight
 import com.missionchurchcooljc.data_android.WebsiteHighlightDAO
+import com.missionchurchcooljc.mcc.network.api.ChurchWebsiteResponse
 import com.missionchurchcooljc.mcc.persistence.MccRoomDatabase
 import com.missionchurchcooljc.mcc.utilities.ABOUT_US_DATA_FILENAME
 import kotlinx.coroutines.runBlocking
+import org.hamcrest.CoreMatchers
 import org.hamcrest.CoreMatchers.equalTo
 import org.junit.After
 import org.junit.Assert.assertThat
@@ -97,8 +99,12 @@ class WebsiteHighlightDaoTest {
 
         val f = context.assets.open(ABOUT_US_DATA_FILENAME)
         val jsonReader = JsonReader(f.reader())
-        val whType = object : TypeToken<List<WebsiteHighlight>>() {}.type
-        val whList: List<WebsiteHighlight> = Gson().fromJson(jsonReader, whType)
+        val whType = object : TypeToken<ChurchWebsiteResponse>() {}.type
+//        val whType = object : TypeToken<List<WebsiteHighlight>>() {}.type
+//        val whList: List<WebsiteHighlight> = Gson().fromJson(jsonReader, whType)
+//        val highlights: List<WebsiteHighlight> = Gson().fromJson(jsonReader, whType)
+        val highlights: ChurchWebsiteResponse = Gson().fromJson(jsonReader, whType)
+        val whList: List<WebsiteHighlight> = highlights.items
         websiteHighlightDAO.insertAll(whList)
 
     }
@@ -116,18 +122,19 @@ class WebsiteHighlightDaoTest {
         assertThat(wListII[1], equalTo(wh2))
 
     }
-    @Test
-    fun checkHighlight3(){
-        val wListII = websiteHighlightDAO.listHighlights()
-        assertThat(wListII[9], equalTo(wh3))
-    }
-    @Test
-    fun checkHighlight4(){
-        val wListII = websiteHighlightDAO.listHighlights()
-        assertThat(wListII[11], equalTo(wh4))
-    }
 
-
+//    @Test
+//    fun checkHighlight3() {
+//        val wListII = websiteHighlightDAO.listHighlights()
+//        val websiteHighlight = wListII //.highlightId = 'ctfw-highlight-52'
+//        assertThat(wListII[9], equalTo(wh3))
+//    }
+//
+//    @Test
+//    fun checkHighlight4() {
+//        val wListII = websiteHighlightDAO.listHighlights()
+//        assertThat(wListII[11], equalTo(wh4))
+//    }
 
 
     // Ensure plant list is sorted by name
@@ -151,7 +158,25 @@ class WebsiteHighlightDaoTest {
 //        assertThat(plantList[1], equalTo(wh2))
 //    }
 
-//    @Test fun testGetPlant() {
+    //    @Test fun testGetPlant() {
 //        assertThat(getValue(websiteHighlightDAO.getHighlight(wh1.websiteHighlightId)), equalTo(wh1))
 //    }
+    @Test
+    fun test_JsonReader() {
+//        TODO: Mock the SeedDatabaseWorker class instead
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        val f = appContext.assets.open(ABOUT_US_DATA_FILENAME)
+        val jsonReader = JsonReader(f.reader())
+        val whType = object : TypeToken<ChurchWebsiteResponse>() {}.type
+//        val highlights: List<WebsiteHighlight> = Gson().fromJson(jsonReader, whType)
+        val highlights: ChurchWebsiteResponse = Gson().fromJson(jsonReader, whType)
+        val whList: List<WebsiteHighlight> = highlights.items
+
+//        assertEquals(jsonReader, JsonReader(f.reader()))
+        whList.forEach { t ->
+            assertThat(t, CoreMatchers.instanceOf(WebsiteHighlight::class.java))
+
+
+        }
+    }
 }
